tests(internal/surfer): improve code coverage
Improve code coverage for internal/surfer by deleting code that has no
effect on tests.

Move gcloud.Generate to surfer.Generate because Generate is the entry
point for the CLI and belongs with the surfer package that owns the CLI
definition.

Add a codecov target of 77% for internal/surfer, with the goal of
reaching 80%.
