+++ codecov.yml
@@ -18,6 +18,14 @@ coverage:
         target: 80% # The target coverage percentage for the project
         threshold: 0% # Allow 0% drop from the target; any drop below target will fail
         if_ci_failed: error # Set the CI status to 'error' if coverage conditions are not met
+      surfer:
+        # TODO(https://github.com/googleapis/librarian/issues/3632): improve
+        # coverage to 80%
+        target: 77%
+        threshold: 0%
+        if_ci_failed: error
+        paths:
+          - internal/surfer/**
     patch: off
 ignore:
   # test packages do not need code coverage
@@ -34,5 +42,3 @@ ignore:
   - internal/legacylibrarian/legacyimages
   # TODO(https://github.com/googleapis/librarian/issues/3633): delete
   - internal/sidekick/sidekick
-  # TODO(https://github.com/googleapis/librarian/issues/3632): improve coverage
-  - internal/surfer/gcloud
