+++ internal/surfer/gcloud/builder.go
@@ -16,7 +16,6 @@ package gcloud
 
 import (
 	"fmt"
-	"slices"
 	"strings"
 
 	"github.com/googleapis/librarian/internal/sidekick/api"
@@ -81,10 +80,6 @@ func NewCommand(method *api.Method, overrides *Config, model *api.API, service *
 		}
 	}
 
-	if method.OperationInfo != nil {
-		cmd.Async = newAsync(method, overrides)
-	}
-
 	return cmd, nil
 }
 
@@ -93,104 +88,47 @@ func NewCommand(method *api.Method, overrides *Config, model *api.API, service *
 //
 // TODO(https://github.com/googleapis/librarian/issues/3412): Refactor to use a dispatch pattern
 // (IsIgnored, IsResourceArg, IsArg) to handle field processing.
-func newArguments(method *api.Method, overrides *Config, model *api.API, service *api.Service) (Arguments, error) {
-	args := Arguments{}
-	if method.InputType == nil {
-		return args, nil
-	}
-
-	// We iterate over each field in the method's request message.
-	for _, field := range method.InputType.Fields {
-		// We process each field and its sub-fields recursively.
-		// TODO(https://github.com/googleapis/librarian/issues/3413): Improve error handling strategy (Error vs Skip) and messaging.
-		if err := addFlattenedParams(field, field.JSONName, &args, overrides, model, service, method); err != nil {
-			return Arguments{}, err
-		}
-	}
-	return args, nil
+func newArguments(_ *api.Method, _ *Config, _ *api.API, _ *api.Service) (Arguments, error) {
+	return Arguments{}, nil
 }
 
-// shouldSkipParam determines if a field should be excluded from the generated command arguments.
-func shouldSkipParam(field *api.Field, method *api.Method) bool {
-	// We don't skip the primary resource field, even if it's named "parent" or "name".
-	if utils.IsPrimaryResource(field, method) {
-		return false
+// newRequest creates the `Request` part of the command definition.
+func newRequest(method *api.Method, overrides *Config, model *api.API) *Request {
+	// TODO(https://github.com/googleapis/librarian/issues/3290): The collection path is partially hardcoded.
+	return &Request{
+		APIVersion: apiVersion(overrides),
+		Collection: []string{fmt.Sprintf("parallelstore.projects.locations.%s", utils.GetPluralResourceNameForMethod(method, model))},
 	}
-
-	// The "parent" field is usually implicit in the command context (handled by the primary resource or hierarchy).
-	if field.Name == "parent" {
-		return true
-	}
-
-	// The "name" field is usually the primary resource identifier, handled separately.
-	if field.Name == "name" {
-		return true
-	}
-
-	// The "update_mask" field is handled automatically by the gcloud framework.
-	if field.Name == "update_mask" {
-		return true
-	}
-
-	// For List methods, standard pagination/filtering arguments are handled by gcloud.
-	if utils.IsList(method) {
-		switch field.Name {
-		case "page_size", "page_token", "filter", "order_by":
-			return true
-		}
-	}
-
-	// Output-only fields are read-only and should not be settable via CLI flags.
-	if slices.Contains(field.Behavior, api.FIELD_BEHAVIOR_OUTPUT_ONLY) {
-		return true
-	}
-
-	// For Update commands, fields marked as IMMUTABLE cannot be changed and should be hidden.
-	if utils.IsUpdate(method) && slices.Contains(field.Behavior, api.FIELD_BEHAVIOR_IMMUTABLE) {
-		return true
-	}
-
-	return false
 }
 
-// addFlattenedParams recursively processes a field and its sub-fields to generate
-// command-line flags. This function identifies primary resources and handles
-// nested messages by "flattening" them into top-level flags.
-func addFlattenedParams(field *api.Field, prefix string, args *Arguments, overrides *Config, model *api.API, service *api.Service, method *api.Method) error {
-	// We check if the field should be skipped entirely.
-	if shouldSkipParam(field, method) {
+// findHelpTextRule finds the help text rule from the config that applies to the current method.
+func findHelpTextRule(method *api.Method, overrides *Config) *HelpTextRule {
+	if overrides.APIs == nil {
 		return nil
 	}
-
-	// We check if the current field represents the primary resource of the command.
-	// This check happens at every level of nesting (e.g., `instance.name`).
-	if utils.IsPrimaryResource(field, method) {
-		param := newPrimaryResourceParam(field, method, model, overrides, service)
-		args.Params = append(args.Params, param)
-		return nil
-	}
-
-	// If the field is a nested message (and not a map), we recurse into its fields.
-	if field.MessageType != nil && !field.Map {
-		for _, f := range field.MessageType.Fields {
-			if err := addFlattenedParams(f, fmt.Sprintf("%s.%s", prefix, f.JSONName), args, overrides, model, service, method); err != nil {
-				return err
+	for _, api := range overrides.APIs {
+		if api.HelpText == nil {
+			continue
+		}
+		for _, rule := range api.HelpText.MethodRules {
+			if rule.Selector == strings.TrimPrefix(method.ID, ".") {
+				return rule
 			}
 		}
-		return nil
 	}
-
-	// If the field is a scalar, map, or enum, we generate a parameter for it.
-	param, err := newParam(field, prefix, overrides, model, service, method)
-	if err != nil {
-		return err
-	}
-	args.Params = append(args.Params, param)
 	return nil
 }
 
+// apiVersion extracts the API version from the configuration.
+func apiVersion(overrides *Config) string {
+	if len(overrides.APIs) > 0 {
+		return overrides.APIs[0].APIVersion
+	}
+	return ""
+}
+
 // newParam creates a single command-line argument (a `Param` struct) from a proto field.
-func newParam(field *api.Field, apiField string, overrides *Config, model *api.API, service *api.Service, method *api.Method) (Param, error) {
+func newParam(field *api.Field, apiField string, overrides *Config, _ *api.API, _ *api.Service, method *api.Method) (Param, error) {
 	// We initialize the Param with the basic information derived from the field.
 	// TODO(https://github.com/googleapis/librarian/issues/3414): Abstract away casing logic in the model.
 	param := Param{
@@ -207,17 +145,8 @@ func newParam(field *api.Field, apiField string, overrides *Config, model *api.A
 
 	// Now we handle the different types of fields.
 	if field.ResourceReference != nil {
-		// If the field is a resource reference (e.g., a field for a network), we
-		// generate a `ResourceSpec` for it. This tells gcloud how to parse the
-		// resource name provided by the user.
-		spec, err := newResourceReferenceSpec(field, model, overrides, service)
-		if err != nil {
-			return Param{}, err
-		}
-		param.ResourceSpec = spec
-		param.ResourceMethodParams = map[string]string{
-			apiField: "{__relative_name__}",
-		}
+		// TODO(https://github.com/googleapis/librarian/issues/3417): Implement resource reference handling.
+		return Param{}, fmt.Errorf("resource references are not yet supported")
 	} else if field.Map {
 		// If the field is a map, we generate a spec for its key-value pairs.
 		param.Repeated = true
@@ -259,189 +188,6 @@ func newParam(field *api.Field, apiField string, overrides *Config, model *api.A
 	return param, nil
 }
 
-// newPrimaryResourceParam creates the main positional resource argument for a command.
-// This is the argument that represents the resource being acted upon (e.g., the instance name).
-func newPrimaryResourceParam(field *api.Field, method *api.Method, model *api.API, _ *Config, service *api.Service) Param {
-	// We first need to get the full resource definition for the method.
-	resource := utils.GetResourceForMethod(method, model)
-	var segments []api.PathSegment
-	// TODO(https://github.com/googleapis/librarian/issues/3415): Support multiple resource patterns and multitype resources.
-	if resource != nil && len(resource.Patterns) > 0 {
-		segments = resource.Patterns[0]
-	}
-
-	// For List methods, the primary resource is the parent of the method's resource.
-	if utils.IsList(method) {
-		segments = utils.GetParentFromSegments(segments)
-	}
-
-	// We determine the singular name of the resource.
-	resourceName := strcase.ToSnake(strings.TrimSuffix(field.Name, "_id"))
-	if field.Name == "name" || utils.IsList(method) {
-		resourceName = utils.GetSingularFromSegments(segments)
-	}
-
-	// We generate a helpful help text.
-	var helpText string
-	switch {
-	case utils.IsCreate(method):
-		helpText = fmt.Sprintf("The %s to create.", resourceName)
-	case utils.IsList(method):
-		helpText = fmt.Sprintf("The project and location for which to retrieve %s information.", utils.GetPluralFromSegments(segments))
-	default:
-		helpText = fmt.Sprintf("The %s to operate on.", resourceName)
-	}
-
-	// We construct the gcloud collection path from the resource's pattern string.
-	collectionPath := utils.GetCollectionPathFromSegments(segments)
-	hostParts := strings.Split(service.DefaultHost, ".")
-	shortServiceName := hostParts[0]
-
-	// We assemble the final `Param` struct.
-	param := Param{
-		HelpText:          helpText,
-		IsPositional:      !utils.IsList(method),
-		IsPrimaryResource: true,
-		Required:          true,
-		ResourceSpec: &ResourceSpec{
-			Name:                  resourceName,
-			PluralName:            utils.GetPluralFromSegments(segments),
-			Collection:            fmt.Sprintf("%s.%s", shortServiceName, collectionPath),
-			DisableAutoCompleters: false,
-			Attributes:            newAttributesFromSegments(segments),
-		},
-	}
-
-	if utils.IsCreate(method) {
-		param.RequestIDField = strcase.ToLowerCamel(field.Name)
-	}
-
-	return param
-}
-
-// newResourceReferenceSpec creates a ResourceSpec for a field that references
-// another resource type (e.g., a `--network` flag).
-func newResourceReferenceSpec(field *api.Field, model *api.API, _ *Config, service *api.Service) (*ResourceSpec, error) {
-	// We iterate through all the resource definitions in the API model to find the
-	// one that matches the type of our resource reference.
-	for _, def := range model.ResourceDefinitions {
-		if def.Type == field.ResourceReference.Type {
-			if len(def.Patterns) == 0 {
-				return nil, fmt.Errorf("resource definition for %q has no patterns", def.Type)
-			}
-			// TODO(https://github.com/googleapis/librarian/issues/3415): Support multiple resource patterns and multitype resources.
-			segments := def.Patterns[0]
-
-			// We determine the plural name, using the explicit `plural` field if available,
-			// and falling back to parsing the pattern otherwise.
-			pluralName := def.Plural
-			if pluralName == "" {
-				pluralName = utils.GetPluralFromSegments(segments)
-			}
-
-			// We determine the singular name from the pattern.
-			name := utils.GetSingularFromSegments(segments)
-
-			// We construct the full gcloud collection path for the referenced resource
-			//assuming the current service is the current command.
-			hostParts := strings.Split(service.DefaultHost, ".")
-			shortServiceName := hostParts[0]
-			baseCollectionPath := utils.GetCollectionPathFromSegments(segments)
-			fullCollectionPath := fmt.Sprintf("%s.%s", shortServiceName, baseCollectionPath)
-
-			// We assemble and return the `ResourceSpec`.
-			return &ResourceSpec{
-				Name:       name,
-				PluralName: pluralName,
-				Collection: fullCollectionPath,
-				// TODO(https://github.com/googleapis/librarian/issues/3416): Investigate and enable auto-completers for referenced resources.
-				DisableAutoCompleters: true,
-				Attributes:            newAttributesFromSegments(segments),
-			}, nil
-		}
-	}
-	return nil, fmt.Errorf("resource definition not found for type %q", field.ResourceReference.Type)
-}
-
-// newAttributesFromSegments parses a structured resource pattern and extracts the attributes
-// that make up the resource's name.
-func newAttributesFromSegments(segments []api.PathSegment) []Attribute {
-	var attributes []Attribute
-
-	// We iterate over the segments of the pattern.
-	for i, part := range segments {
-		// A variable segment is enclosed in curly braces.
-		if part.Variable == nil {
-			continue
-		}
-
-		// The `attribute_name` is the name of the variable (e.g., "project").
-		if len(part.Variable.FieldPath) == 0 {
-			continue
-		}
-		name := part.Variable.FieldPath[len(part.Variable.FieldPath)-1]
-		var parameterName string
-
-		// The `parameter_name` is derived from the preceding literal segment
-		// (e.g., "projects" -> "projectsId"). This is a gcloud convention.
-		if i > 0 && segments[i-1].Literal != nil {
-			parameterName = *segments[i-1].Literal + "Id"
-		} else {
-			// This is a fallback for the unlikely case that a pattern starts with a variable.
-			parameterName = name + "sId"
-		}
-
-		attr := Attribute{
-			AttributeName: name,
-			ParameterName: parameterName,
-			Help:          fmt.Sprintf("The %s id of the {resource} resource.", name),
-		}
-
-		// If the attribute is a project, we add the standard gcloud property fallback,
-		// so users don't have to specify `--project` if it's already configured.
-		if name == "project" {
-			attr.Property = "core/project"
-		}
-		attributes = append(attributes, attr)
-	}
-	return attributes
-}
-
-// newRequest creates the `Request` part of the command definition.
-func newRequest(method *api.Method, overrides *Config, model *api.API) *Request {
-	// TODO(https://github.com/googleapis/librarian/issues/3290): The collection path is partially hardcoded.
-	return &Request{
-		APIVersion: apiVersion(overrides),
-		Collection: []string{fmt.Sprintf("parallelstore.projects.locations.%s", utils.GetPluralResourceNameForMethod(method, model))},
-	}
-}
-
-// newAsync creates the `Async` part of the command definition for long-running operations.
-func newAsync(_ *api.Method, _ *Config) *Async {
-	return &Async{
-		// TODO(https://github.com/googleapis/librarian/issues/3290): The collection path is partially hardcoded.
-		Collection: []string{"parallelstore.projects.locations.operations"},
-	}
-}
-
-// findHelpTextRule finds the help text rule from the config that applies to the current method.
-func findHelpTextRule(method *api.Method, overrides *Config) *HelpTextRule {
-	if overrides.APIs == nil {
-		return nil
-	}
-	for _, api := range overrides.APIs {
-		if api.HelpText == nil {
-			continue
-		}
-		for _, rule := range api.HelpText.MethodRules {
-			if rule.Selector == strings.TrimPrefix(method.ID, ".") {
-				return rule
-			}
-		}
-	}
-	return nil
-}
-
 // findFieldHelpTextRule finds the help text rule from the config that applies to the current field.
 func findFieldHelpTextRule(field *api.Field, overrides *Config) *HelpTextRule {
 	if overrides.APIs == nil {
@@ -458,12 +204,4 @@ func findFieldHelpTextRule(field *api.Field, overrides *Config) *HelpTextRule {
 		}
 	}
 	return nil
-}
-
-// apiVersion extracts the API version from the configuration.
-func apiVersion(overrides *Config) string {
-	if len(overrides.APIs) > 0 {
-		return overrides.APIs[0].APIVersion
-	}
-	return ""
 }
